ai_incubation_platform/data-agent-connector/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
aiomysql>=0.2.0
asyncpg>=0.28.0
psycopg2-binary>=2.9.0
pymongo>=4.6.0
motor>=3.3.0
redis>=5.0.0
elasticsearch>=8.0.0
aiosqlite>=0.19.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
python-json-logger>=2.0.7
anthropic>=0.18.0
google-cloud-bigquery>=3.0.0
grpcio>=1.50.0
grpcio-status>=1.50.0
clickhouse-connect>=0.6.0
databricks-sql-connector>=3.0.0

# 非结构化数据连接器依赖
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.0.0
python-pptx>=0.6.23
openpyxl>=3.1.0
pandas>=2.0.0
Pillow>=10.0.0
paddlepaddle>=2.5.0
paddleocr>=2.7.0
pytesseract>=0.3.10
playwright>=1.40.0
beautifulsoup4>=4.12.0
aioimaplib>=1.0.2
lxml>=5.0.0

# P5 数据智能功能依赖 - LlamaIndex 框架
llama-index>=0.10.0
llama-index-vector-stores-chroma>=0.1.0
llama-index-core>=0.10.0
llama-index-llms-openai>=0.1.0
llama-index-embeddings-openai>=0.1.0
chromadb>=0.4.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# 孵化器统一 Agent 运行时（DeerFlow 2.0）— 见 PLATFORM_AGENT_STANDARD.md
-e ../deerflow-integration
